requirements.txt
asgiref==3.2.5
certifi==2020.6.20
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
psycopg2==2.8.5
pytz==2019.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
